json v1.1.0 JSON.Parser.Unicode

Implements a JSON Unicode Parser for Bitstring values

Functions

parse(json)

parses a valid chain of escaped unicode and returns the string representation, plus the remainder of the string

Examples

iex> JSON.Parser.parse ""
{:error, :unexpected_end_of_buffer}

iex> JSON.Parser.parse "face0ff"
{:error, {:unexpected_token, "face0ff"}}

iex> JSON.Parser.parse "-hello"
{:error, {:unexpected_token, "-hello"}}